fix: cleanup account details component
